+++ packages/flame/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## [Next]
+ - Add `ButtonComponent` backed by two `PositionComponent`s
+
 ## [1.0.0-releasecandidate.18]
  - Forcing portrait and landscape mode is now supported on web
  - Fixed margin calculation in `HudMarginComponent` when using a viewport